+++ tests/test_init.py
@@ -1,4 +1,3 @@
-from typing import Any
 from unittest.mock import ANY
 
 from bleak.backends.device import BLEDevice
@@ -36,20 +35,14 @@ def test_create_scanner():
 def test_create_remote_scanner():
     connector = HaBluetoothConnector(MockBleakClient, "any", lambda: True)
 
-    def callback(data: Any) -> None:
-        pass
-
-    scanner = BaseHaRemoteScanner("any", "any", callback, connector, True)
+    scanner = BaseHaRemoteScanner("any", "any", connector, True)
     assert isinstance(scanner, BaseHaRemoteScanner)
 
 
 def test__async_on_advertisement():
     connector = HaBluetoothConnector(MockBleakClient, "any", lambda: True)
 
-    def callback(data: Any) -> None:
-        pass
-
-    scanner = BaseHaRemoteScanner("any", "any", callback, connector, True)
+    scanner = BaseHaRemoteScanner("any", "any", connector, True)
     details = scanner._details | {}
     scanner._async_on_advertisement(
         "AA:BB:CC:DD:EE:FF",
@@ -105,10 +98,5 @@ def test__async_on_advertisement():
 
 
 def test_create_ha_scanner():
-    def callback(data: Any) -> None:
-        pass
-
-    scanner = HaScanner(
-        BluetoothScanningMode.ACTIVE, "hci0", "AA:BB:CC:DD:EE:FF", callback
-    )
+    scanner = HaScanner(BluetoothScanningMode.ACTIVE, "hci0", "AA:BB:CC:DD:EE:FF")
     assert isinstance(scanner, HaScanner)
